(WIP) Broken attempt at filtering macOS undefined symbols / giving the same output as nm ..
Broken because:
All the stuff about class n_type_* and class n_stab* is wrong, this stuff is
just struct ranlib or struct ranlib_64 as can be seen in:

/opt/MacOSX10.14.sdk/usr/include/mach-o/ranlib.h

and struct ranlib can be read about here:

http://mirror.informatimago.com/next/developer.apple.com/documentation/DeveloperTools/Conceptual/MachORuntime/8rt_file_format/chapter_10_section_33.html

Note: I have not yet seen a ranlib_64 in the wild, they are identified via:

.. so currently the code cannot handle this at all.

Using lief on actual objects (i.e. all the files after the index file)
is also a possibility, but not needing lief would be preferable for speed
reasons. The variable to look for in that case though is content

Pinging @msarahan, if you can fix this up that would be excellent!