fix pip build isolation
The environment variable PIP_NO_BUILD_ISOLATION is set to 'False' in the
build environment.  Previously this variable was not set in the environment
due to it having a falsey value. It is now set to the string 'False' so that it
has a truethy value in Python which results in in being included in the
environment file.

closes #3318