(WIP) Attempting to allow using build tools to download patches
.. probably want to roll this in with the effort to allow submodule
patches instead and tackle that later. Some patches can be applied
with git and some can only be applied with {m2-,}patch and we need
to know which is which.

This commit is far from correct. Its current purpose is just to serve
as bookmarks for where in the code we likely need to change things to
make this work.