On ASCII locales, allow UTF-8 text in scripts
The noarch check for binary scripts uses the locale preferred encoding to guess if the script is binary or text based.

On some CI systems, (e.g. gitlab continuous integration using the `conda/miniconda3` docker image), the default encoding is ascii (expressed as "ANSI_X3.4-1968").

In general, in any system with an ascii default encoding conda assumes "if it is not ascii then it is binary". I argue it makes more sense to say "if it is not utf-8 then it is binary". Since ascii is a subset of UTF-8, we will reduce the false positives with this change.

Other encodings are not affected by this.