Update launcher sources to handle debug shebangs
in debug mode (not sure what). It's easy enough to handle this.

We special-case this to avoid expensive lookups and make it in-line
with what happens when there is no shebang (release builds).

The launcher.c source code expects to be built twice, once for each
of debug and release and defines PYTHON_EXECUTABLE differently for
each. This isn't really what we want though, esp. once the ABIs are
exactly the same between debug and release pythons.